requirements.txt
Flask==3.0.3 
pytest==8.3.4
pytest-flask==1.3.0
pythondotenv==1.0.1
pyjwt==2.10.1